✨ Site Version 2.0.0
I know I should describe the changes in the commit message, but I'm
lazy. I'll add an emoji to make it look like I put in some effort. 😅

Seriously though, this commit is a big one. I've made a lot of changes
to the site. I'm super excited to see how it looks now.